LPD-3866 Moving test to other class
