+++ modules/apps/segments/segments-service/src/main/java/com/liferay/segments/internal/exportimport/data/handler/SegmentsExperienceStagedModelDataHandler.java
@@ -167,7 +167,8 @@ public class SegmentsExperienceStagedModelDataHandler
 		}
 
 		if ((existingSegmentsExperience == null) ||
-						!portletDataContext.isDataStrategyMirror()) {
+				!portletDataContext.isDataStrategyMirror()) {
+	
 			importedSegmentsExperience = _stagedModelRepository.addStagedModel(
 				portletDataContext, importedSegmentsExperience);
 		}
