+++ modules/apps/segments/segments-test/src/testIntegration/java/com/liferay/segments/exportimport/data/handler/test/SegmentsExperienceStagedModelDataHandlerTest.java
@@ -95,10 +95,6 @@ public class SegmentsExperienceStagedModelDataHandlerTest
 			importedSegmentsExperience.getPriority());
 	}
 
-	@Override
-	protected void testFileUploadSelfreference(
-		
-	}
 	private long _classPK;
 
 	@Inject
